apps/api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic>=2.5.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
typer>=0.9.0
orjson>=3.9.10
jsonschema>=4.20.0
python-magic>=0.4.27
pillow>=10.1.0

# Development
ruff>=0.1.0
black>=23.0.0
